center the auditionees at the bottom

+++ db/migrate/20250909132929_drop_audition_sessions.rb
@@ -1,6 +1,10 @@
 class DropAuditionSessions < ActiveRecord::Migration[8.0]
   def change
-    drop_table :auditions
-    drop_table :audition_sessions
+    if table_exists?(:auditions)
+      drop_table :auditions
+    end
+    if table_exists?(:audition_sessions)
+      drop_table :audition_sessions
+    end
   end
 end
